requirements.txt
absl-py==1.0.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==8.1.3
Flask==3.0.0
Flask-WTF==1.2.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.48.2
gunicorn==20.0.4
h5py==3.10.0
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
Markdown==3.1.1
MarkupSafe==2.1.1
numpy==1.23.5
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.13.0
tensorflow-cpu==2.8.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==3.0.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
